fix(tests): corrige 27 fallos preexistentes y contratos de tipos
- deps.py: HTTPBearer(auto_error=False) + raise 401 explícito al faltar token (era 403)
- schemas: amount en SaleEntry/ExpenseEntry/Product usa Decimal en lugar de float
- test_endpoints: patch paths actualizados a chat_orchestrator; test_rate_limit no requiere mock de AgentCEO
- test_sales/expenses: _TODAY usa date.today() en lugar de fecha hardcodeada fuera de ventana 30d
- test_workspace: 4 tests marcados skip (OAuth Workspace reemplazado por MCP)
- test_fase0/fase1b: _get_db_conn skipea si DATABASE_URL es SQLite; token cipher mockea get_settings

+++ backend/app/schemas/product.py
@@ -15,8 +15,8 @@ class ProductResponse(BaseModel):
     sku: str | None
     description: str | None
     category: str | None
-    sale_price_ars: float
-    unit_cost_ars: float | None
+    sale_price_ars: Decimal
+    unit_cost_ars: Decimal | None
     stock_units: int
     low_stock_threshold_units: int
     is_active: bool
